Cutting queries in index and tags.
Tags now stores a compressed version of tags in the images table so as to quicken page load.  Things are still stored in the tags database for search though.  

New set_tags() function that pulls the tags for an image from the tags table and then sticks them into the tags column

git-svn-id: file:///home/oestrich/tmp/img/trunk@1137 ea04b438-0a61-0410-93d0-9bc465a3bab0